requirements.txt
# Malaria Cell Classification# Malaria Cell Classification - Core Requirements

# Production-ready ML pipeline for malaria parasite detection

# Python 3.9+# Deep Learning

torch>=2.0.1

# === Core Deep Learning ===torchvision>=0.15.2

torch>=2.0.1,<3.0.0

torchvision>=0.15.2,<1.0.0# Data handling

numpy>=1.24.3

# === Data Processing ===pillow>=10.0.0

numpy>=1.24.3,<2.0.0pyyaml>=6.0.1

pillow>=10.0.0,<11.0.0

pyyaml>=6.0.1,<7.0.0# ML Pipeline & Versioning

tqdm>=4.65.0  # Progress barsdvc>=3.0.0



# === Development & Visualization ===# Visualization & Development

matplotlib>=3.7.0,<4.0.0matplotlib>=3.7.0

seaborn>=0.12.0,<1.0.0seaborn>=0.12.0

jupyter>=1.0.0

# === Web API (Production Deployment) ===

# Uncomment these for deploying the Flask API# Optional: Uncomment for deployment

# flask>=3.0.0,<4.0.0# flask>=2.3.3

# gunicorn>=21.2.0,<22.0.0# gunicorn>=21.2.0
# werkzeug>=3.0.0,<4.0.0

# === Testing & Code Quality (Development) ===
# Uncomment for development and testing
# pytest>=7.4.0,<8.0.0
# pytest-cov>=4.1.0,<5.0.0
# black>=23.7.0,<24.0.0
# flake8>=6.1.0,<7.0.0
# mypy>=1.5.0,<2.0.0

# === ML Pipeline Tools (Optional) ===
# Uncomment if using data versioning or experiment tracking
# dvc>=3.0.0,<4.0.0
# mlflow>=2.7.0,<3.0.0
